feat(compiler): Working on watching
